requirements.txt
# pip install -r requirements.txt


# base ----------------------------------------
Cython
matplotlib==3.3.4
numpy==1.20.1
opencv-python==4.5.1.48
Pillow==8.1.2
PyYAML>=5.3.1
scipy==1.6.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0
Shapely==1.7.1
cycler==0.10.0
kiwisolver==1.3.1
pyparsing==2.4.7
python-dateutil==2.8.1

# logging -------------------------------------
# wandb

# plotting ------------------------------------
seaborn>=0.11.0
pandas

# export --------------------------------------
# coremltools>=4.1
# onnx>=1.8.1
# scikit-learn==0.19.2  # for coreml quantization

# extras --------------------------------------
thop  # FLOPS computation
pycocotools>=2.0  # COCO mAP

python-telegram-bot==13.0

stockfish==3.13.0
chess==1.4.0
cairosvg